requirement.txt
flask
pillow
pix2tex[cli]
